% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_map_onto_plate.R
\name{plate_map_onto_plate}
\alias{plate_map_onto_plate}
\alias{adding_technical_replicate_suffix}
\title{map a single plate_map to a read promega_plate}
\usage{
adding_technical_replicate_suffix(mapped_plate)
}
\arguments{
\item{mapped_plate}{the dataframe of values which columns are assigned but may have repeated columns}
}
\value{
a dataframe where all values have _techrep{n} suffixed, prevents having repeated column names.
}
\description{
Where there are repeated experiment columns, give them a suffix to indicate the number of technical replicates,
and more importantly prevent automatic suffix addition in R to prevent unexpected results.
This acts as a daughter function of plate_map_onto_plate but also can act as a stand alone function.
}
