% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_in_one.R
\name{final_func}
\alias{final_func}
\title{Strings all functions together.}
\usage{
final_func(neut_xlsx_path, sheetname = "Results", rotation_deg_needed = 0)
}
\arguments{
\item{rotation_deg_needed}{degrees that dataframe needs to be rotated, default =0. Feeds into rotate.R}

\item{df}{dataframe to normalise, where left most column is negative control, and second left column is positive control.}

\item{control_neg}{The column which negative/cell control are located (column number from left, integer,default=1)}

\item{control_pos}{The column which positive/cell control are located (column number from left, integer,default=2)}
}
\value{
A rotated, normalised dataframe where negative and positive control columns are removed.
}
\description{
Strings all functions together.
}
