% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_clustering.R
\name{k_clustering}
\alias{k_clustering}
\title{install.packages("Ckmeans.1d.dp")
select k-cluster mean that is closest to the mean of your first and last rows without outliers}
\usage{
k_clustering(
  control_column,
  lim,
  type = c("negative", "positive"),
  na.rm = TRUE
)
}
\description{
install.packages("Ckmeans.1d.dp")
select k-cluster mean that is closest to the mean of your first and last rows without outliers
}
