% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_without_outliers.R
\name{mean_without_outliers}
\alias{mean_without_outliers}
\title{Remove the outliers from your first and last rows (upper and lower bounds) and find the mean
Define numeric_vector}
\usage{
mean_without_outliers(numeric_vector, num_sd = 2)
}
\description{
Remove the outliers from your first and last rows (upper and lower bounds) and find the mean
Define numeric_vector
}
