% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_map.R
\docType{data}
\name{plate_map_functions}
\alias{plate_map_functions}
\alias{expected_columns}
\alias{validate_plate_map}
\alias{generate_plate_map}
\title{Read in and validates, or generate PlateMap following strict formatting as Example_PlateMap.xlsx}
\format{
An object of class \code{character} of length 11.
}
\usage{
expected_columns
}
\arguments{
\item{plateMap_file}{PlateMap excel file name}

\item{sheet}{the sheet which the wanted PlateMap resides on in the plateMap_file}

\item{directory_of_files}{the directory_of_files that is iterated through to look for Promega output files to generated output_plateMap_file.}

\item{output_plateMap_file}{filepath where the generated PlateMap will be located.}
}
\description{
Read in and validates, or generate PlateMap following strict formatting as Example_PlateMap.xlsx
}
\keyword{datasets}
