% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_promega_plate_xlsx.R
\name{read_promega_plate_excel}
\alias{read_promega_plate_excel}
\title{Reading in Promega plate data, from the excel sheet format.}
\usage{
read_promega_plate_excel(input_promega_excel_file_path, sheetname = "Results")
}
\arguments{
\item{input_promega_excel_file_path}{Path to the input file}
}
\value{
A 8x12 data.frame with rownames and column headers
}
\description{
This function loads the Promega system generated excel sheet in a 8x12 matrix
whilst keeping the column and row names.
}
