% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_in_one.R
\name{final_func}
\alias{final_func}
\title{Strings all functions together.}
\usage{
final_func(
  neut_xlsx_path,
  control_neg_column = c(1),
  control_pos_column = c(2),
  sheetname = "Results",
  rotation_deg_needed = 0
)
}
\arguments{
\item{neut_xlsx_path}{a file to normalise, where by default negative control is left most column, and positive control is second left column.}

\item{control_neg_column}{The column which negative/cell control are located (column number from left, integer,default=1)}

\item{control_pos_column}{The column which positive/cell control are located (column number from left, integer,default=2)}

\item{sheetname}{the sheet of the .xlsx file to read, default for promega excel file is "Results"}

\item{rotation_deg_needed}{degrees that dataframe needs to be rotated, default =0. Feeds into rotate.R}
}
\value{
A rotated, normalised dataframe where negative and positive control columns are removed.
}
\description{
Strings all functions together.
}
