% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_map.R
\docType{data}
\name{plate_map_functions}
\alias{plate_map_functions}
\alias{expected_columns}
\alias{validate_plate_map}
\alias{generate_plate_map}
\alias{read_validate_plate_map}
\alias{find_promega_plate_paths}
\alias{get_read_date}
\title{Read in and validates, or generate PlateMap following strict formatting as Example_PlateMap.xlsx}
\format{
An object of class \code{character} of length 11.
}
\usage{
expected_columns

validate_plate_map(df, plateMap_file, read_mode = "default")

read_validate_plate_map(plateMap_file, sheet = "Sheet1")

find_promega_plate_paths(directory_of_files)

get_read_date(promega_excel_file)

generate_plate_map(
  directory_of_files = "Validation/",
  output_plateMap_file = "Validation/generated_platemap.xlsx",
  output_mode = "return"
)
}
\arguments{
\item{df}{The read in platemap dataframe}

\item{plateMap_file}{the filepath used to generate df.
The validate function is required in both reading in and writing out to make sure the expected_columns are included.}

\item{read_mode}{"default" which silently adds missing column or "debug" which is invoked which provides warning of missing/unexpected columns.}

\item{sheet}{the sheet which the wanted PlateMap resides on in the plateMap_file}

\item{directory_of_files}{the directory_of_files that is iterated through to look for Promega output files to generated output_plateMap_file.}

\item{promega_excel_file}{A promega .xlsx file which has intact metadata (first sheet).}

\item{output_plateMap_file}{filepath where the generated PlateMap will be located.}

\item{output_mode}{"return" will return into a dataframe variable, "write" will write to excel file defined by output_plateMap_file}
}
\value{
A dataframe, or an excel file with Wells prepared and two additional columns. 1) The file path of the plates, and 2) the file creation date which corresponds to read date.
}
\description{
Validates the PlateMap by checking if there are unexpected or missing columns, and if there are missing columns, adding them.
}
\section{general}{

}

\section{read-created-plate-map}{

}

\section{generate-plate-map}{





}

\keyword{datasets}
