% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_map_onto_plate.R
\name{plateMap_map_onto_promega_read}
\alias{plateMap_map_onto_promega_read}
\title{finding the plate using the path provided on the plate_map, and mapping the appropriate columns.}
\usage{
plateMap_map_onto_promega_read(plate_map, promega_read_values = "")
}
\arguments{
\item{plate_map}{The platemap with strict definitions validated by plate_map.R}

\item{promega_read_values}{promega_plate_df generated from read_promega_plate_xlsx.R , if not specified, the plate_map column promega_plate_path will be searched}
}
\value{
normalized values with \if{html}{\out{<Group>}}, <dilution_or_concentration>, <dilution_serie> from plate_map
}
\description{
finding the plate using the path provided on the plate_map, and mapping the appropriate columns.
}
