% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Reading in Promega plate data, from the excel sheet format.}
\usage{
rotate(x)
}
\arguments{
\item{x}{dataframe to rotate}
}
\value{
A dataframe rotated 90 degrees clockwise
}
\description{
This function rotates a data.frame of any size by 90 degrees.
}
